incorporating notification changes
